Backend/requirements.txt
# Django and related dependencies
Django==3.2.18  # Compatible with djongo; adjust if needed
djangorestframework==3.14.0  # For API development
django-cors-headers==4.2.0  # For handling CORS in frontend integration

# Database dependencies (choose one based on your database)
mysqlclient==2.1.1  # MySQL (if using MySQL)
djongo==1.3.6  # MongoDB (if using MongoDB)

# Machine Learning & Image Processing (For AI-powered ripeness detection)
opencv-python==4.7.0.72  # OpenCV for image processing
numpy==1.24.3  # Numerical computing
torch==2.0.1  # PyTorch for deep learning
ultralytics==8.0.176  # YOLO model package

# Background Task Management
celery==5.3.4  # Asynchronous task management
redis==5.0.1  # Celery broker (if using Redis)

# Environment & Utility Libraries
python-dotenv==1.0.0  # For environment variable management
requests==2.31.0  # For making HTTP requests
Pillow==10.0.1  # For handling image uploads
geopy==2.3.0  # Geolocation for smart shipment tracking
drf-yasg==1.21.5  # API documentation (Swagger)

# Testing Libraries
pytest-django==4.5.2  # Django testing
pytest==7.4.0  # General testing framework
